requirements.txt
Flask
flask_cors
protobuf
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf       
python-docx    
openpyxl
pandas
python-pptx